Updates related to slangpy documentation updates (#21)
* Update URLs in examples/README.md

Updates URLs in examples/README.md
Reorders examples to match documentation order

* Add return_type example

* Revise Compute Shader tutorial opening note

This change addresses an item in issue 339 in the slangpy repo.

* Update README.md

* test_examples.py: Add return_type coverage

* test_examples.py: Flush expected before assert

The subsequent assert may terminate the process before the file created
by numpy.savez() is flushed to disk, so add an explicit flush. (Issue
seen on Windows)

* Updates to match documentation